+++ libs/block-mesh-common/Cargo.toml
@@ -7,3 +7,4 @@ authors.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+tracing-subscriber = { workspace = true, features = ["env-filter", "time", "json"] }
