+++ libs/block-mesh-common/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod constants;
+pub mod tracing;
