+++ libs/test-proxy-endpoint/Cargo.toml
@@ -17,3 +17,4 @@ hyper-util = { workspace = true }
 rustc-hash = { workspace = true }
 http-body-util = { workspace = true }
 bytes = { workspace = true }
+block-mesh-common = { path = "../block-mesh-common" }
