+++ libs/test-proxy-master/Cargo.toml
@@ -20,3 +20,4 @@ once_cell = { workspace = true }
 futures-util = { workspace = true }
 anyhow = { workspace = true }
 httparse = { workspace = true }
+block-mesh-common = { path = "../block-mesh-common" }
